test(delivery): replace WaitForPickupUseCase tests with AssignCourierUseCase tests
Replace wait-for-pickup.spec.ts with assign-courier.spec.ts. Add courier repository and courier creation to test setup. Update test cases to verify courier assignment with courierId parameter and ASSIGNED status instead of WAITING_PICKUP. Update error assertions to use CannotAssignCourierError
